+++ web/singleplayer/singleplayer.js
@@ -23,15 +23,14 @@ async function preparePage() {
     let uuid = getCookie("uuid");
     if (uuid == "" || uuid == null) {
         await register();
-        await updatePage();
+        updatePage();
     } else {
         // some cookie exists, check if cookie is valid
         let status = await (wasm_bindgen.get_request("api/registered")) + '';
-        console.log(status);
         switch (status) {
             case 'false':// uuid is invalid
                 await register();            
-                await updatePage();
+                updatePage();
                 break;
             case 'playing':// player is still playing the game
                 updatePage();
@@ -59,7 +58,6 @@ async function updatePage() {
     updateWord();
     updateGuessedChars();
     updateLives();
-    updateImage();
 }
 
 /**
@@ -68,27 +66,67 @@ async function updatePage() {
 async function updateWord() {
     let word = await (wasm_bindgen.get_request("api/game_string"));
     document.getElementById("word").innerHTML = word;
-    document.getElementById("word-placeholder").hidden = true;
     document.getElementById("word").hidden = false;
+    document.getElementById("word-placeholder").hidden = true;
 }
 
 /**
  * Update the guessed characters
  */
 async function updateGuessedChars() {
-
+    let guessed_chars = await (wasm_bindgen.get_request("api/guessed_letters"));
+    document.getElementById("guessed-letters").innerHTML = guessed_chars;
+    document.getElementById("guessed-letters").hidden = false;
+    document.getElementById("guessed-letters-placeholder").hidden = true;
 }
 
 /**
- * Update the lives
+ * Update the lives.
+ * Automatically updates the image.
  */
 async function updateLives() {
-
+    let lives = await (wasm_bindgen.get_request("api/lives"));
+    document.getElementById("lives-left").innerHTML = lives;
+    document.getElementById("lives-left").hidden = false;
+    document.getElementById("lives-left-placeholder").hidden = true;
+    updateImage(lives);
 }
 
 /**
  * Update the image
+ * @param {int} lives_left - The amount of lives left
  */
-async function updateImage() {
+async function updateImage(lives_left) {
+    document.getElementById("image").src = "pictures/" + Math.abs(10 - Number(lives_left)) + ".jpg"
+    document.getElementById("image").hidden = false;
+    document.getElementById("image-placeholder").hidden = true;
+}
 
+/**
+ * Sends the content of the text field to the server to make a guess
+ */
+async function guess() {
+    var response = await postData('api/guess', document.getElementById("user-input").value);
+    console.log(response);
+    switch (response) {
+      case 1: 
+        updatePage();
+        //gameEnd(true);
+        break;
+      case 2: 
+        updatePage();
+        break;
+      case 3: 
+        updatePage();
+        break;
+      case 4: 
+        updatePage();
+        //gameEnd(false);
+        break;
+      case 5:
+        alert("This character was already submitted");
+        document.getElementById("user-input").value = "";
+        break;
+    }
+    document.getElementById("user-input").value = "";
 }