+++ changelog.md
@@ -2,5 +2,9 @@
 
 ## v2.0.0
 
+### New features
+- Single player mode
+- It is now supported to submit words
+
 ### Removed features
 - Removed multiplayer (will perhaps be added again at a later point of time)